Move initial view timings in dedicated files
