+++ packages/rum-core/src/domain/rumEventsCollection/view/viewCollection.ts
@@ -79,22 +79,22 @@ function processViewUpdate(
       frustration: {
         count: view.eventCounts.frustrationCount,
       },
-      cumulative_layout_shift: view.cumulativeLayoutShift,
-      first_byte: toServerDuration(view.timings.firstByte),
-      dom_complete: toServerDuration(view.timings.domComplete),
-      dom_content_loaded: toServerDuration(view.timings.domContentLoaded),
-      dom_interactive: toServerDuration(view.timings.domInteractive),
+      cumulative_layout_shift: view.metrics.cumulativeLayoutShift,
+      first_byte: toServerDuration(view.initialViewMetrics.firstByte),
+      dom_complete: toServerDuration(view.initialViewMetrics.domComplete),
+      dom_content_loaded: toServerDuration(view.initialViewMetrics.domContentLoaded),
+      dom_interactive: toServerDuration(view.initialViewMetrics.domInteractive),
       error: {
         count: view.eventCounts.errorCount,
       },
-      first_contentful_paint: toServerDuration(view.timings.firstContentfulPaint),
-      first_input_delay: toServerDuration(view.timings.firstInputDelay),
-      first_input_time: toServerDuration(view.timings.firstInputTime),
+      first_contentful_paint: toServerDuration(view.initialViewMetrics.firstContentfulPaint),
+      first_input_delay: toServerDuration(view.initialViewMetrics.firstInputDelay),
+      first_input_time: toServerDuration(view.initialViewMetrics.firstInputTime),
       is_active: view.isActive,
       name: view.name,
-      largest_contentful_paint: toServerDuration(view.timings.largestContentfulPaint),
-      load_event: toServerDuration(view.timings.loadEvent),
-      loading_time: discardNegativeDuration(toServerDuration(view.loadingTime)),
+      largest_contentful_paint: toServerDuration(view.initialViewMetrics.largestContentfulPaint),
+      load_event: toServerDuration(view.initialViewMetrics.loadEvent),
+      loading_time: discardNegativeDuration(toServerDuration(view.metrics.loadingTime)),
       loading_type: view.loadingType,
       long_task: {
         count: view.eventCounts.longTaskCount,
