+++ packages/rum-core/src/domain/rumEventsCollection/view/viewMetrics/trackCumulativeLayoutShift.spec.ts
@@ -53,7 +53,7 @@ describe('trackCumulativeLayoutShift', () => {
     const { getViewUpdate, getViewUpdateCount } = viewTest
 
     expect(getViewUpdateCount()).toEqual(1)
-    expect(getViewUpdate(0).cumulativeLayoutShift).toBe(0)
+    expect(getViewUpdate(0).metrics.cumulativeLayoutShift).toBe(0)
   })
 
   it('should be initialized to undefined if layout-shift is not supported', () => {
@@ -62,7 +62,7 @@ describe('trackCumulativeLayoutShift', () => {
     const { getViewUpdate, getViewUpdateCount } = viewTest
 
     expect(getViewUpdateCount()).toEqual(1)
-    expect(getViewUpdate(0).cumulativeLayoutShift).toBe(undefined)
+    expect(getViewUpdate(0).metrics.cumulativeLayoutShift).toBe(undefined)
   })
 
   it('should accumulate layout shift values for the first session window', () => {
@@ -74,7 +74,7 @@ describe('trackCumulativeLayoutShift', () => {
     clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
 
     expect(getViewUpdateCount()).toEqual(2)
-    expect(getViewUpdate(1).cumulativeLayoutShift).toBe(0.3)
+    expect(getViewUpdate(1).metrics.cumulativeLayoutShift).toBe(0.3)
   })
 
   it('should round the cumulative layout shift value to 4 decimals', () => {
@@ -86,7 +86,7 @@ describe('trackCumulativeLayoutShift', () => {
     clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
 
     expect(getViewUpdateCount()).toEqual(2)
-    expect(getViewUpdate(1).cumulativeLayoutShift).toBe(2.3457)
+    expect(getViewUpdate(1).metrics.cumulativeLayoutShift).toBe(2.3457)
   })
 
   it('should ignore entries with recent input', () => {
@@ -98,7 +98,7 @@ describe('trackCumulativeLayoutShift', () => {
     clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
 
     expect(getViewUpdateCount()).toEqual(1)
-    expect(getViewUpdate(0).cumulativeLayoutShift).toBe(0)
+    expect(getViewUpdate(0).metrics.cumulativeLayoutShift).toBe(0)
   })
 
   it('should create a new session window if the gap is more than 1 second', () => {
@@ -114,7 +114,7 @@ describe('trackCumulativeLayoutShift', () => {
 
     clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
     expect(getViewUpdateCount()).toEqual(2)
-    expect(getViewUpdate(1).cumulativeLayoutShift).toBe(0.3)
+    expect(getViewUpdate(1).metrics.cumulativeLayoutShift).toBe(0.3)
   })
 
   it('should create a new session window if the current session window is more than 5 second', () => {
@@ -127,7 +127,7 @@ describe('trackCumulativeLayoutShift', () => {
     } // window 1: 0.5 | window 2: 0.1
     clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
     expect(getViewUpdateCount()).toEqual(3)
-    expect(getViewUpdate(2).cumulativeLayoutShift).toBe(0.5)
+    expect(getViewUpdate(2).metrics.cumulativeLayoutShift).toBe(0.5)
   })
 
   it('should get the max value sessions', () => {
@@ -148,6 +148,6 @@ describe('trackCumulativeLayoutShift', () => {
 
     clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
     expect(getViewUpdateCount()).toEqual(3)
-    expect(getViewUpdate(2).cumulativeLayoutShift).toBe(0.5)
+    expect(getViewUpdate(2).metrics.cumulativeLayoutShift).toBe(0.5)
   })
 })
