+++ packages/rum-core/src/domain/rumEventsCollection/view/viewMetrics/trackFirstHidden.spec.ts
@@ -1,7 +1,7 @@
 import type { RelativeTime } from '@datadog/browser-core'
 import { DOM_EVENT } from '@datadog/browser-core'
 import { createNewEvent, restorePageVisibility, setPageVisibility } from '@datadog/browser-core/test'
-import type { RumConfiguration } from '../../configuration'
+import type { RumConfiguration } from '../../../configuration'
 import { resetFirstHidden, trackFirstHidden } from './trackFirstHidden'
 
 describe('trackFirstHidden', () => {
